packages/discord-bot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"discord_bot" = "src"}
packages = ["discord_bot"]

[project]
name = "ryuuko-discord-bot"
version = "2.0.0"
authors = [
  { name="Zang Vũ", email="zvwgvx@polydevs.uk" },
]
description = "Discord client for the Ryuuko Chatbot ecosystem."
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "Pillow>=10.0.0"
]
